+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 22.09.0
-* Events are now recorded with an internal ID.
 * Fixed a race condition bug where a recorded event would have the wrong user properties in the drill database on the server. Now event queue is emptied (formed into a request) before recording any user profile changes.
+* Events are now recorded with an internal ID.
+* Added a variant of "esWriteCachesToPersistence" that has a callback to indicate if any writes were performed
+* When calling "esWriteCachesToPersistence", no write to storage is performed if the same data would be written
 
 ## 22.06.3
 * Expanded feedback widget functionality. Added ability to use rating widgets.